src/third_party/blink/web_tests/external/wpt/mixed-content/gen/top.http-rp/opt-in/sharedworker-import-data/cross-http.keep-scheme.https-expected.txt

This is a testharness.js-based test.
FAIL Mixed-Content: Expects blocked for sharedworker-import-data to cross-http origin and keep-scheme redirection from https context. assert_equals: The resource request should be 'blocked'. expected "blocked" but got "allowed"
Harness: the test ran to completion.
